.insert(), attributes improved
The foundation for this class is solid now, and the .insert() method works properly. There's still a need to tie the end of this function to a verification of the node: if there's more elements than the maximum allowed, then the DegreeOverflow Exception shall be raised, where the proper repairs are to be made.