.insert(), .find() created, overall improvements
There has been quite some changes around here. First, BTree.insert() has been created. It uses the other method created, .find(), to see if the element to be inserted already exists. If it does, nothing is changed; otherwise, since the .find() method returns the page which is closest to the right place, BTree.insert() acts upon the returned page by inserting the element in it (through BTree.Page.insert() method). In case the Page.insert() method, after insertion of the item, raises an DegreeOverflowError(), the BTree.insert() method will handle it by calling .promoteRoot(), which will separate the page into two ones.

There's way more to be said than it has been, but I think this will suffice. It is pertinent to say that basic testing has proven these methods to be working; however, further testing is necessary to be sure.