.index() created, self.descendent_pages changed to []
Following Giovanne's advice on the employment of self.descendent_pages, instead of creating an array full of None objects, I decided to add the page elements along the way. In addition, Page.index() was created to give to Page classes the same functionality that list objects do when retrieving the index of inserted objects.