Breadth-first search fixed (I think)! And other stuff
It is recursive as well. Not that Gabriela's implementation wasn't working; it's just that I wanted to try it with a different spin. Might replace it with Gabriela's implementation later, depending if it's really working or not.

In addition, more verifications have been added to return meaningful errors to the user.