Bunch of changes. Quite something left to be done.
I tried to create a linked_components property prior to the strongly_linked_components; however, I didn't succeed in my approach. I'll decide later whether to implement it or just go to the strongly-linked part of it. The implementation of the transposed graph, however, is complete and seemingly functional.

In addition, all lists which are return (either as values of keys in dictionaries or as the result of calling properties or definitions) have been converted to tuples, to accentuate the immutability of the results obtained.

The notebook has been improved and extended to reflect and test the module's capabilities.