The Graph implementation seems to be done! And documented!
Of course, testing is never enough and there might be some corner cases somewhere; however, these graphs seem to be pretty much final.

There are many new methods and properties – most of them in the Graph superclass –, but the main goal of their appearance (although far from the only one) was to create the Prim, Kruskal and Dijkstra algorithms in the Graph object. During this process, many ideas came up about modifying and flexibilizing the existing methods to make their employment in possible in new situations and contexts.

I'll try to ask Gabriela to take a look at it tomorrow. I still need to clean some thing up around here.