Broken, old version of old Graphs module removed
This happened as I was trying to understand Gabriela's code: I created a single module, attempting to unify the classes under a single umbrella and changing other things in the process. This resulted in an unconsistent implementation; as such, it must be removed.

Let's just hope that I didn't mess up with the original implementation... whatever, it's in another branch anyway.