Eccentricities, distances matrix, radius and diameter implemented
Nobody asked for it, but I got the impression it will be useful next semester. At least, I did it to hone the implementation. This is how I roll, I think y'all got it this far.

Oh, btw, I tested it on two examples of the slides, but given that I fixed a lot of inconsistencies throughout the existing code, I wouldn't be surprised if I introduced a couple of bugs along the way.